requirements.txt
# ── Data & Core ──────────────────────────────────────────────
yfinance==0.2.54
pandas-datareader==0.10.0
pandas==2.2.3
numpy==1.26.4
requests==2.32.3

# ── Technical Analysis ────────────────────────────────────────
ta==0.11.0

# ── Database ─────────────────────────────────────────────────
psycopg2-binary==2.9.10
sqlalchemy==2.0.36

# ── Environment ───────────────────────────────────────────────
python-dotenv==1.0.1

# ── Machine Learning ─────────────────────────────────────────
scikit-learn==1.5.2
xgboost==2.1.3
lightgbm==4.5.0

# ── Visualization ─────────────────────────────────────────────
matplotlib==3.9.3
seaborn==0.13.2

# ── Notebooks (VS Code maneja Jupyter nativamente) ───────────
ipykernel==6.29.5
